+++ apps/meteor/app/lib/server/functions/deleteRoom.ts
@@ -2,6 +2,7 @@ import { Messages, Rooms, Subscriptions } from '@rocket.chat/models';
 
 import { callbacks } from '../../../../lib/callbacks';
 import { FileUpload } from '../../../file-upload/server';
+import { notifyOnRoomChangedById } from '../lib/notifyListener';
 
 export const deleteRoom = async function (rid: string): Promise<void> {
 	await FileUpload.removeFilesByRoomId(rid);
@@ -11,4 +12,6 @@ export const deleteRoom = async function (rid: string): Promise<void> {
 	await FileUpload.getStore('Avatars').deleteByRoomId(rid);
 	await callbacks.run('afterDeleteRoom', rid);
 	await Rooms.removeById(rid);
+
+	void notifyOnRoomChangedById(rid, 'removed');
 };