+++ apps/meteor/client/components/message/content/attachments/file/ImageAttachment.tsx
@@ -2,8 +2,9 @@ import type { ImageAttachmentProps } from '@rocket.chat/core-typings';
 import { useMediaUrl } from '@rocket.chat/ui-contexts';
 import React from 'react';
 
+import MarkdownText from '../../../../MarkdownText';
 import MessageCollapsible from '../../../MessageCollapsible';
-import AttachmentDescription from '../structure/AttachmentDescription';
+import MessageContentBody from '../../../MessageContentBody';
 import AttachmentImage from '../structure/AttachmentImage';
 import { useLoadImage } from './hooks/useLoadImage';
 
@@ -28,7 +29,7 @@ const ImageAttachment = ({
 
 	return (
 		<>
-			<AttachmentDescription description={description} descriptionMd={descriptionMd} />
+			{descriptionMd ? <MessageContentBody md={descriptionMd} /> : <MarkdownText parseEmoji content={description} />}
 			<MessageCollapsible title={title} hasDownload={hasDownload} link={getURL(link || url)} size={size} isCollapsed={collapsed}>
 				<AttachmentImage
 					{...imageDimensions}